requirements.txt
amqp==1.4.9
asgiref==3.5.2
billiard==3.3.0.23
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-adminlte3==0.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.3.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumberslite==8.12.53
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.2.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
